% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_C.R
\name{cal_C}
\alias{cal_C}
\title{Estimate \code{C}.}
\usage{
cal_C(Y, x, Z_uc, Z_w, B, trimmed_data, na_covariates)
}
\arguments{
\item{Y}{A numeric vector; realizations of the outcome variable.}

\item{x}{A numeric vector; realizations of the treatment variable.}

\item{Z_uc}{A numeric matrix; realizations of the control covariates.}

\item{Z_w}{A numeric matrix; realizations of the interacted covariates.}

\item{B}{A numeric matrix; covariance matrix of the demeaned \code{Z_uc}.}

\item{trimmed_data}{A list; output from \code{trim_data()}, used if the long regression has \code{NA} coefficients.}

\item{na_covariates}{A character vector; names of interaction covariates with \code{NA} coefficients with the
\code{Z_w} prefix removed obtained from running \code{find_long_NAs()}.}
}
\value{
A list containing the square of the estimated \code{C} and its associated standard error, and CI.
}
\description{
Estimate \code{C}, the heterogeneity bound, from the data. If the long regression includes
covariates with \code{NA} coefficients, use the trimmed data instead.
}
