% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt_het_bound.R
\name{calculate_lambda_range}
\alias{calculate_lambda_range}
\title{Calculate an appropriate range for finding the optimal lambda value.}
\usage{
calculate_lambda_range(
  res_prop,
  B,
  resid_Z_w_Z_uc,
  x,
  short_wghts,
  scaling = 2
)
}
\arguments{
\item{res_prop}{A numeric vector; residuals from projecting \code{x} onto \code{Z_uc}.}

\item{B}{A numeric matrix; the covariance matrix of the demeaned \code{Z_uc}.}

\item{resid_Z_w_Z_uc}{A numeric matrix; residualized \code{Z_w} after projection onto \code{Z_uc}.}

\item{x}{A numeric vector; realizations corresponding to the treatment variable.}

\item{short_wghts}{A numeric vector; weights used in the short regression specification.}

\item{scaling}{A numeric scalar; a scaling factor for the maximum \eqn{\lambda} value. Defaults to 2.}
}
\value{
A numeric vector; containing the minimum and maximum value of the lambda range.
}
\description{
Determines a feasible search interval for the regularization parameter \eqn{\lambda} used in penalized propensity score regression.
The minimum value of this interval is always set to zero.
The maximum is initialized using the largest \eqn{\lambda} value returned by \code{glmnet} because it recovers the fully regularized
short regression coefficient estimates. The maximum \eqn{\lambda} is then adaptively increased by 2 until the \eqn{\ell_2}-distance
between the weight vectors of this penalized propensity score regression and that of the fully penalized propensity score regression
falls below a pre-specified tolerance.
}
