% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regulaTE.R
\name{check_valid_inputs}
\alias{check_valid_inputs}
\title{Verify user inputs to regulaTE() are valid.}
\usage{
check_valid_inputs(
  outcome,
  treatment,
  covariates,
  parameter,
  C,
  se,
  cluster,
  sig_level,
  df_corr,
  digits,
  confounders,
  trimmed_data
)
}
\arguments{
\item{outcome}{A numeric vector; realizations of the outcome variable.}

\item{treatment}{A numeric vector; realizations of the treatment variable.}

\item{covariates}{A numeric matrix; realizations of the control covariates.}

\item{parameter}{A string; the parameter of interest, which must be \code{"ATE"}, \code{"ATT"}, or \code{"ATU"}. Defaults to \code{"ATE"}.}

\item{C}{A numeric scalar or vector; contains non-negative values specifying the heterogeneity bound or
\code{NA} if one wishes the program to estimate \code{C} using the data.}

\item{se}{A string; either \code{"het"}, or \code{"hom"}. \code{"hom"} computes homoskedastic standard errors. \code{"het"} computes
heteroskedastic-robust standard errors. If \code{"het"} is chosen and \code{cluster} is specified, then it computes clustered
heteroskedastic-robust standard errors.}

\item{cluster}{A vector; cluster identifier for computing clustered standard errors. Defaults to \code{NULL}. If \code{NULL}, non-clustered standard errors are used.}

\item{sig_level}{A numeric vector; significance level for confidence intervals. Defaults to 0.05.}

\item{df_corr}{indicating whether to apply degrees-of-freedom correction. Defaults to \code{TRUE}.}

\item{digits}{An integer vector; number of decimal places to use when formmatting numeric values. Defaults to 4.}

\item{confounders}{A numeric matrix; realizations of the confounders. Defaults to \code{NULL}.
If \code{NULL}, the function will use the covariates as confounders.}

\item{trimmed_data}{A list; contains the matrices/vectors for the outcome (\code{Y}),
treatment indicator (\code{x}), and interaction variables (\code{Z_w}), plus either
covariates (\code{Z_uc}) or confounders (\code{Z_conf}). Used when trimming is necessary due to lack of
overlap and the user wants to specify the trimmed data directly. Defaults to \code{NULL},
in which case the function trims the data automatically.}
}
\value{
None if inputs are valid; otherwise, the function stops the program.
}
\description{
Verify user inputs to regulaTE() are valid.
}
