% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{check_valid_plot_inputs}
\alias{check_valid_plot_inputs}
\title{Checks if inputs into plot_regulaTE() are valid.}
\usage{
check_valid_plot_inputs(estimates, y_axis_title, title = NULL)
}
\arguments{
\item{estimates}{A data frame; \code{estimates_df} from regulaTE().}

\item{y_axis_title}{A character string; the title for the y-axis. Defaults to "Outcome."}

\item{title}{A character string; the tile for the graph. Defaults to "Estimates under Heterogeneity Bound for \if{html}{\out{<Parameter>}}"}
}
\value{
None if inputs are valid; otherwise, the function stops the program.
}
\description{
Checks if inputs into plot_regulaTE() are valid.
}
