% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/console_output.R
\name{clean_estimates_table}
\alias{clean_estimates_table}
\title{Clean parameter estimates table for formatted output.}
\usage{
clean_estimates_table(table, parameter, digits, sig_level)
}
\arguments{
\item{table}{A data frame; contains the point estimates, standard errors, and confidence intervals corresponding to the parameter estimates
from \code{opt_het_bound()}.}

\item{parameter}{A character vector; the parameter of interest, which must be one of \code{"ATE"}, \code{"ATT"}, or \code{"ATU"}.}

\item{digits}{An integer vector; number of decimal places to use when formmatting numeric values.}

\item{sig_level}{A numeric vector; significance level for confidence intervals.}
}
\value{
A data frame; contains the point estimates, standard errors, and confidence intervals corresponding
to the parameter estimates from \code{opt_het_bound()} formalized appropriately for console output.
}
\description{
Produces a cleaned version of the estimates table from \code{opt_het_bound()}, formatted for display via \code{print_pretty_table()}.
}
