% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/console_output.R
\name{console_output}
\alias{console_output}
\title{Console output for regulaTE().}
\usage{
console_output(
  output,
  estimates_table,
  parameter,
  digits,
  sig_level,
  C,
  multi = FALSE
)
}
\arguments{
\item{output}{A list; output from \code{opt_het_bound()}}

\item{estimates_table}{A data frame; produced by \code{produce_estimates_table()}.}

\item{parameter}{A character string; the parameter of interest, which must be one of \code{"ATE"}, \code{"ATT"}, or \code{"ATU"}.}

\item{digits}{An numeric vector; number of decimal places to display in the output.}

\item{sig_level}{A numeric vector; significance level for confidence intervals.}

\item{C}{A numeric vector; non-negative values specifying the heterogeneity bound or \code{NA} if one wishes the program to estimate \code{C} using the data.}

\item{multi}{A logical vector; whether the output is based on multiple \code{C} values. Defaults to \code{FALSE}.}
}
\value{
None. Only prints to console.
}
\description{
Prints formatted console output for results from \code{opt_het_bound()}.
}
