% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_computations.R
\name{covariate_comps}
\alias{covariate_comps}
\title{Preliminary covariate computations}
\usage{
covariate_comps(x, Z_uc, parameter = "ATE")
}
\arguments{
\item{x}{A numeric vector; realizations of the treatment variable.}

\item{Z_uc}{A numeric matrix; realizations of the control covariates.}

\item{parameter}{A character string; Specifies which treatment effect ("ATE", "ATU", or "ATT") to compute. Defaults to "ATE".}
}
\value{
A list containing the following components:
\itemize{
\item \code{parameter}: A character string; the specified treatment effect ("ATE", "ATT", or "ATU").
\item \code{n}: An integer; number of observations (rows).
\item \code{x}: A numeric vector;realizations of the treatment variable.
\item \code{Z_uc}: A numeric matrix; realizations of the control covariates.
\item \code{Z_tilde}: A numeric matrix; the demeaned version of \code{Z_uc} according to the specified \code{parameter}.
\item \code{Z_w}: A numeric matrix; realizations cof the interacted covariates.
}
}
\description{
Performs preliminary computations only dependent on the covariates. Computes demeaned covariates and interaction terms baese on the specified treatment effect parameter.
}
