% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimates_table.R
\name{cr_se_loop}
\alias{cr_se_loop}
\title{Cluster-robust standard error computation}
\usage{
cr_se_loop(wghts, residuals, cluster, num_covar, df_corr = TRUE)
}
\arguments{
\item{wghts}{A numeric vector; weights applied to the residuals.}

\item{residuals}{A numeric vector; residuals.}

\item{cluster}{A vector; cluster identifiers for each obsevation.}

\item{num_covar}{An integer vector; the number of covariates in the regression model.}

\item{df_corr}{A logical vector; if \code{TRUE}, applies a finite-sample correction to the standard error calculation.}
}
\value{
A numeric scalar corresponding to the cluster-robust standard error.
}
\description{
Computes cluster-robust, heteroskedasticity-robust standard errors for a linear combination of residuals and weights,
with optional finite-sample degrees-of-freedom correction. It is thus equivalent to the \code{HC1} estimator (default in Stata).
}
