% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_estimates.R
\name{determine_residuals}
\alias{determine_residuals}
\title{Long or ridge regression residuals}
\usage{
determine_residuals(long_NAs, long_reg, ridgereg)
}
\arguments{
\item{long_NAs}{A character vector; names of interaction covariates with \code{NA} coefficients, with the \code{Z_w} prefix removed. Possibly empty.}

\item{long_reg}{An \code{lm} object; the long regression object, \code{lm(Y ~ -1 + x + Z_uc + Z_w)}.}

\item{ridgereg}{A \code{glmnet} object; the ridge regression object obtained from running \code{run_long_ridge()}.}
}
\value{
A list containing the following components:
\itemize{
\item \code{residuals}: A numeric vector; residuals, either from the long regression or ridge long regression.
\item \code{df}: A numeric vector; degrees of freedom corresponding to the long regression or ridge long regression.
}
}
\description{
Returns residuals, either from the long regression or ridge long regression,
depending on whether the long regression contains any \code{NA} coefficients.
}
