% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_estimates.R
\name{find_long_NAs}
\alias{find_long_NAs}
\title{\code{NA} coefficients in long regression detector}
\usage{
find_long_NAs(long_reg)
}
\arguments{
\item{long_reg}{An \code{lm} object; the long regression object, \code{lm(Y ~ -1 + x + Z_conf + Z_w)}.}
}
\value{
A character vector; containing the names of interaction covariates with \code{NA} coefficients with the
\code{Z_w} prefix removed. If there are no such covariates, returns an empty character vector.
}
\description{
Returns the names of the \code{NA} coefficients in the interaction covariates of a given long regression: these
are the coefficients contributing to lack of overlap and hence a long regression with \code{NA} coefficients.
}
