% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimates_table.R
\name{generate_estimates_table}
\alias{generate_estimates_table}
\title{Generate a table of estimates for each regression specifcation}
\usage{
generate_estimates_table(
  opt_results,
  reg_estimates,
  sigma,
  alpha,
  zalpha,
  cva,
  max_bias_short,
  max_bias_trim,
  cluster,
  df_corr
)
}
\arguments{
\item{opt_results}{A list; output from running \code{opt_lam()}.}

\item{reg_estimates}{A list; components obtained from running \code{reg_estimates()}.}

\item{sigma}{A numeric vector; residual standard deviation used to compute robust SEs when clustering is not specified.}

\item{alpha}{A numeric vector; between \code{0} and \code{1}; significance level for CIs.}

\item{zalpha}{A numeric vector; standard normal quantile for two-sided CI.}

\item{cva}{A numeric vector; quantile from folded normal distribution used for bias correction of the short regression.}

\item{max_bias_short}{A numeric vector; the maximum bias of the short regression.}

\item{max_bias_trim}{A numeric vector; the maximum bias of the trimmed long regression}

\item{cluster}{A vector; optional cluster identifiers for each obsevation. If \code{NULL}, standard errors are not clustered.}

\item{df_corr}{A logical vector; indicating whether to apply degrees-of-freedom correction. Defaults to \code{TRUE}.}
}
\value{
A numeric matrix; each row corresponds to a regression specification and includes
point estimates (PE), standard errors (SE), confidence intervals (CI LB and CI UB), and maximum bias, and critical value.
}
\description{
Generate a table of estimates for each regression specifcation
}
