% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt_het_bound.R
\name{opt_het_bound}
\alias{opt_het_bound}
\title{Inference on treatment effects using a bound on heterogeneity}
\usage{
opt_het_bound(
  Y,
  reg_estimates,
  cluster,
  C,
  alpha = 0.05,
  df_corr = TRUE,
  scaling = 2
)
}
\arguments{
\item{Y}{A numeric vector; outcome variable.}

\item{reg_estimates}{A list; output from \code{reg_estimates()}.}

\item{cluster}{A vector; cluster identifier for computing clustered standard errors. If \code{NULL}, non-clustered standard errors are used.}

\item{C}{A numeric vector; heterogeneity bound. If \code{NULL}, estimated using \code{cal_C()}.}

\item{alpha}{A numeric vector; significance level for confidence intervals. Defaults to 0.05.}

\item{df_corr}{A logical; indicating whether to apply degrees-of-freedom correction. Defaults to \code{TRUE}.}

\item{scaling}{A numeric scalar; a scaling factor for the maximum \eqn{\lambda} value. Defaults to 2.}
}
\value{
A list with the following components:
\itemize{
\item \code{estimates_table}: A numeric matrix; estimates for Opt, Short, Short BC, and (Trimmed) Long specs produced by \code{generate_estimates_table()}.
\item \code{c_table}: A numeric matrix or \code{NA}; If \code{C} is estimated, a matrix containing PE, SE, and CI for \code{C} obtained from \code{generate_estimates_table()}; otherwise \code{NA}.
\item \code{trimming_table}: A numeric matrix or \code{NA}; If trimming of the long regression is required, a matrix containing details of the trimming process obtained from \code{generate_estimates_table()}; otherwise \code{NA}.
\item \code{lambda}: A numeric vector; value of the regularization parameter that minimizes CI length.
\item \code{C_orig}:  A numeric vector or \code{NULL}; the user-supplied value of \code{C} or \code{NULL} if estimated using \code{cal_C()}.
\item \code{samp_size}: A numeric vector; number of observations in the data.
\item \code{num_covar_short}: A numeric vector; number of covariates used in the short regression.
\item \code{num_covar_long}: A numeric vector; number of covariates used in the long regression.
\item \code{lind_wt}: A numeric vector; the Lindeberg weight.
\item \code{max_bias_short}: A numeric vector; the maximum bias of the short regression.
}
}
\description{
Estimates the treatment effects (ATE, ATT, or ATU) using four different regressions: Opt, Short, Short BC, (Trimmed) Long.
}
