% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_regulaTE}
\alias{plot_regulaTE}
\title{Plots multiple estimates from regulaTE() with varying \code{C} values.}
\usage{
plot_regulaTE(
  estimates,
  y_axis_title = "Outcome",
  title = NULL,
  show_long = FALSE
)
}
\arguments{
\item{estimates}{A data frame; \code{estimates_df} from \code{regulaTE()}.}

\item{y_axis_title}{A character string; the title for the y-axis. Defaults to "Outcome."}

\item{title}{A character string; the title for the plot. Defaults to "Estimates under Heterogeneity Bound for \if{html}{\out{<Parameter>}}"}

\item{show_long}{A logical; whether to show the long regression results in the plot. Defaults to FALSE.}
}
\value{
Note. Only prints a plot.
}
\description{
Plots multiple estimates from regulaTE() with varying \code{C} values.
}
