% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/console_output.R
\name{print_pretty_table}
\alias{print_pretty_table}
\title{Prints a well-formatted table to the console}
\usage{
print_pretty_table(df, digits, pad = 2)
}
\arguments{
\item{df}{A data frame; contains parameter of \code{C} estimates from \code{produce_estimates_table()} or \code{produce_c_table()}.}

\item{digits}{An integer vector; number of decimal places to use when formmatting numeric values.}

\item{pad}{An integer vector; number of spaces between the columns. Defaults to 2.}
}
\value{
None. Only prints to console.
}
\description{
Prints a well-formatted table to the console for a data frame of parameter or \code{C} estimates.
}
