% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/console_output.R
\name{produce_c_table}
\alias{produce_c_table}
\title{Produces \code{C} estimates table for formatted output.}
\usage{
produce_c_table(table, digits, sig_level)
}
\arguments{
\item{table}{A data frame; contains the estimates.}

\item{digits}{An integer vector; number of decimal places to use when formmatting numeric values.}

\item{sig_level}{A numeric vector; significance level for confidence intervals.}
}
\value{
A data frame; contains the point estimate, standard error, and confidence interval corresponding to the
estimated \eqn{\vec{C}} from \code{opt_het_bound()} formalized appropriately for console output.
}
\description{
Produces a cleaned data frame containing the estimated \eqn{C^2}, standard error, and confidence interval, formatted for display via \code{print_pretty_table()}.
}
