% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_estimates.R
\name{reg_estimates}
\alias{reg_estimates}
\title{Short, long, and ridge regressions}
\usage{
reg_estimates(Y, x, covariate_comps, Z_conf = NULL, trimmed_data = NULL)
}
\arguments{
\item{Y}{A numeric vector; realizations of the outcome variable.}

\item{x}{A numeric vector; realizations of the treatment variable.}

\item{covariate_comps}{A list; covariate-related components returned by \code{covariate_comps()}.}

\item{Z_conf}{A numeric matrix; realizations of the confounder covariates. Defaults to \code{Z_uc} if \code{NULL}.}

\item{trimmed_data}{A list; contains the matrices/vectors for the outcome (\code{Y}),
treatment indicator (\code{x}), and interaction variables (\code{Z_w}), plus either
covariates (\code{Z_uc}) or confounders (\code{Z_conf}). Used when trimming is necessary due to lack of
overlap and the user wants to specify the trimmed data directly. Defaults to \code{NULL},
in which case the function trims the data automatically.}
}
\value{
A list with the following components:
\itemize{
\item \code{multi_results}: A list; contains (if applicable):
\itemize{
\item \code{ridgereg}: A \code{glmnet} object; the ridge regression object (or \code{NULL} if unused).
\item \code{trim_reg}: An \code{lm} object; the trimmed long regression object.
\item \code{res_long_prop_trim}: A numeric vector; residuals from projecting \code{x} on trimmed covariates.
\item \code{long_wghts_trim}: A numeric vector; weights from trimmed long regression.
\item \code{long_wghts_trim_full}: A numeric vector; weights from trimmed long regression, but with zeros at trimmed rows.
\item \code{rows_to_delete}: An integer vector; trimmed row indices.
}
\item \code{covariate_comps}: A list; the original, unchanged input to the function.
\item \code{long_reg}: A \code{lm} object; \code{Y ~ - 1 + x + Z_conf + Z_w}.
\item \code{long_NAs}: An integer vector; indices for coefficients dropped in the long regression.
\item \code{num_covar_long}: An integer; number of covariates in the long regression.
\item \code{num_covar_trim}: An integer; number of covariates in the trimmed regression.
\item \code{shortreg}: An \code{lm} object; \code{Y ~ - 1 + x + Z_conf}.
\item \code{residuals_df}: A list; obtained from \code{determine_residuals()} with:
\itemize{
\item \code{res}: A numeric vector; residual, either from the long or ridge long regression.
\item \code{df}: An integer; residual degrees of freedom
}
\item \code{sigma}: A numeric vector; estimated residual standard deviation.
\item \code{wghts}: A numeric matrix; the matrix cross-product of \code{short_wghts} and \code{Z_w}.
\item \code{res_long_prop}: A numeric vector; residuals from projecting \code{x} onto \code{Z_conf} and \code{Z_w}.
\item \code{long_wghts}: A numeric vector; weights used in the long regression specification.
\item \code{num_covar_long}: An integer; number of covariates in the long regression.
\item \code{res_prop}: A numeric vector; residuals from projecting \code{x} onto \code{Z_conf}.
\item \code{short_wghts}: A numeric vector; weights used in the short regression specification.
\item \code{short_prop}: An \code{lm} object; corresponds to the regression of \code{x} on \code{Z_conf}.
\item \code{resid_Z_w_Z_uc}: A numeric matrix; residualized \code{Z_w} after projection onto \code{Z_conf}.
\item \code{Z_w_resid_Z_w_Z_uc}: A numeric matrix; matrix cross-product of residualized \code{Z_w}.
\item \code{Z_w_res_prop}: A numeric vector; matrix cross-product of residualized \code{Z_w} with \code{res_prop}.
\item \code{B}: A numeric matrix; the covariance matrix of the demeaned \code{Z_uc}.
\item \code{inv_B}: A numeric matrix; the inverse of \code{B}.
}
}
\description{
Estimates long and short regressions, as well as a trimmed long and ridge regression
(via \code{run_long_ridge()}) if \code{NA} coefficients exist.
}
