% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regulaTE.R
\name{regulaTE}
\alias{regulaTE}
\title{Function performing inference under a heterogeneity bound that end users call.}
\usage{
regulaTE(
  outcome,
  treatment,
  covariates,
  confounders = NULL,
  parameter = "ATE",
  C = NA,
  se = "hom",
  cluster = NULL,
  sig_level = 0.95,
  df_corr = TRUE,
  digits = 4,
  trimmed_data = NULL,
  scaling = 2
)
}
\arguments{
\item{outcome}{A numeric vector; realizations of the outcome variable.}

\item{treatment}{A numeric vector; realizations of the treatment variable.}

\item{covariates}{A numeric matrix; realizations of the control covariates.}

\item{confounders}{A numeric matrix; realizations of the confounders. Defaults to \code{NULL}.
If \code{NULL}, the function will use the covariates as confounders.}

\item{parameter}{A string; the parameter of interest, which must be \code{"ATE"}, \code{"ATT"}, or \code{"ATU"}. Defaults to \code{"ATE"}.}

\item{C}{A numeric scalar or vector; contains non-negative values specifying the heterogeneity bound or
\code{NA} if one wishes the program to estimate \code{C} using the data.}

\item{se}{A string; either \code{"het"}, or \code{"hom"}. \code{"hom"} computes homoskedastic standard errors. \code{"het"} computes
heteroskedastic-robust standard errors. If \code{"het"} is chosen and \code{cluster} is specified, then it computes clustered
heteroskedastic-robust standard errors.}

\item{cluster}{A vector; cluster identifier for computing clustered standard errors. Defaults to \code{NULL}. If \code{NULL}, non-clustered standard errors are used.}

\item{sig_level}{A numeric vector; significance level for confidence intervals. Defaults to 0.05.}

\item{df_corr}{indicating whether to apply degrees-of-freedom correction. Defaults to \code{TRUE}.}

\item{digits}{An integer vector; number of decimal places to use when formmatting numeric values. Defaults to 4.}

\item{trimmed_data}{A list; contains the matrices/vectors for the outcome (\code{Y}),
treatment indicator (\code{x}), and interaction variables (\code{Z_w}), plus either
covariates (\code{Z_uc}) or confounders (\code{Z_conf}). Used when trimming is necessary due to lack of
overlap and the user wants to specify the trimmed data directly. Defaults to \code{NULL},
in which case the function trims the data automatically.}

\item{scaling}{A numeric scalar; a scaling factor for the maximum \eqn{\lambda} value. Defaults to 2.}
}
\value{
A list of two data frames:
\itemize{
\item \code{summary}: A data frame; contains the sample size, number of covariates in the
short and long regressions, and the Lindeberg weight.
\item \code{estimates}: A data frame; contains point estimates, standard errors, and confidence
intervals for each specification (Opt, Short, Short BC, and Long).
}
}
\description{
Function performing inference under a heterogeneity bound that end users call.
}
