% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_estimates.R
\name{run_long_ridge}
\alias{run_long_ridge}
\title{Ridge long regression}
\usage{
run_long_ridge(Y, x, Z_conf, Z_w, B)
}
\arguments{
\item{Y}{A numeric vector; realizations of the outcome variable.}

\item{x}{A numeric vector; realizations of the treatment variable.}

\item{Z_conf}{A numeric matrix; realizations of the confounder covariates.}

\item{Z_w}{A numeric matrix; realizations cof the interacted covariates.}

\item{B}{A numeric matrix; covariance matrix of the demeaned \code{Z_uc}.}
}
\value{
A list containing the following components:
\itemize{
\item \code{ridge}: An \code{glmnet} object; fitted ridge regression model, an object of class \code{glmnet}.
\item \code{residuals}: A numeric vector; residuals.
\item \code{df}: A numeric vector; the effective degrees of freedom for the ridge regression model,
computed as \code{n - sum(d^2 / (d^2 + lambda))}, where \code{d} are the singular
values of the design matrix \code{X} and \code{lambda} is the optimal penalty
determined by cross-validation.
}
}
\description{
Runs a ridge long regression with a penalty determined by 10-fold cross-validation and computes
the residuals and effective degrees of freedom. Used when the long regression has \code{NA} coefficients.
}
