% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_estimates.R
\name{sqrt_mat}
\alias{sqrt_mat}
\title{Matrix square root}
\usage{
sqrt_mat(A)
}
\arguments{
\item{A}{A numeric matrix.}
}
\value{
A numeric matrix; representing the (approximate) principal square root of \code{A}.
}
\description{
Computes the (approximate) principal square root \eqn{B} of a matrix \eqn{A},
so that \eqn{BB \approx A}, via Schur decomposition.
}
