% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trimming.R
\name{trim_data}
\alias{trim_data}
\title{Trim data associated with limited overlap.}
\usage{
trim_data(data, long_NAs, parameter)
}
\arguments{
\item{data}{A matrix; \code{cbind(Y, x, Z_uc)}, where \code{Y}, \code{x}, and \code{Z_uc} correspond to realizations of the outcome variable,
the treatment variable, and the covariate variables, respectively.}

\item{long_NAs}{A character vector; names of interaction covariates with \code{NA} coefficients, with the \code{Z_w} prefix removed. Possibly empty.}

\item{parameter}{A string; the parameter of interest, which must be \code{"ATE"}, \code{"ATT"}, or \code{"ATU"}. Defaults to \code{"ATE"}.}
}
\value{
A list containing the following components:
\itemize{
\item \code{Y}: A numeric vector; realizations of the outcome variable after trimming.
\item \code{x}: A numeric vector; realizations of the treatment variable after trimming.
\item \code{Z_uc}: A numeric matrix; realizations of the covariates after trimming.
\item \code{Z_w}: A numeric matrix; realizations of the interacted covariates after trimming.
\item \code{Z_tilde}: A numeric matrix; demeaned version of \code{Z_uc}, based on \code{parameter}.
\item \code{B_trim}: A numeric matrix; covariance matrix of the demeaned \code{Z_uc}.
\item \code{trimming_table}: A numeric matrix; 2-row summary of trimming across rows and columns.
\item \code{trimmed_row_indices}: A numeric vector; indices of rows removed due to limited overlap.
}
Here "trimmed" means trimmed according to the procedures in \code{trim_na_rows()} and \code{trim_na_cols()}.
}
\description{
Trims rows and columns that cause an undefined long regression. This function should be used only when \code{Z_conf = Z_uc} (i.e., no confounders).
}
