% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trimming.R
\name{trim_na_cols}
\alias{trim_na_cols}
\title{Removes NA-causing columns (columns associated limited overlap) in dataset.}
\usage{
trim_na_cols(data, na_covariates)
}
\arguments{
\item{data}{A matrix; \code{cbind(Y, x, Z_uc)}, where \code{Y}, \code{x}, and \code{Z_uc} correspond to realizations of the outcome variable,
the treatment variable, and the covariate variables, respectively.}

\item{na_covariates}{A character vector; contains the names of interaction covariates with \code{NA} coefficients with the
\code{Z_w} prefix removed obtained from running \code{find_long_NAs()}.}
}
\value{
A list containing the following components:
\itemize{
\item \code{data}: The dataset with NA columns removed.
\item \code{cols_before}: A numeric scalar; equal to the number of columns in the data before trimming.
\item \code{cols_after}: A numeric saclar; equal to the number of columns in the data after trimming.
\item \code{cols_trimmed}: A numeric scalar; equal to the number of columns trimmed.
\item \code{cols_trimmed_percentage}: A numeric scalar; equal to the percentage of how many columns were trimmed.
}
}
\description{
Helper function for \code{trim_data()}. Removes ("trims") all columns whose names are included in \code{na_covariates}.
}
