% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trimming.R
\name{trim_na_rows}
\alias{trim_na_rows}
\title{Helper function for \code{trim_data()}. Removes ("trims") all rows that have a value of 1 for any of the covariates
passed through in \code{na_covariates}.}
\usage{
trim_na_rows(data, na_covariates)
}
\arguments{
\item{data}{A matrix; \code{cbind(Y, x, Z_uc)}, where \code{Y}, \code{x}, and \code{Z_uc} correspond to realizations of the outcome variable,
the treatment variable, and the covariate variables, respectively.}

\item{na_covariates}{A character vector; contains the names of interaction covariates with \code{NA} coefficients with the
\code{Z_w} prefix removed obtained from running \code{find_long_NAs()}.}
}
\value{
A list containing the following components:
\itemize{
\item \code{data}: The dataset with NA rows removed.
\item \code{deleted_rows}: A numeric vector; contains the indices of the rows that were deleted.
\item \code{rows_before}: A numeric scalar; equal to the number of rows in the data before trimming.
\item \code{rows_after}: A numeric saclar; equal to the number of rows in the data after trimming.
\item \code{rows_trimmed}: A numeric scalar; equal to the number of rows trimmed.
\item \code{rows_trimmed_percentage}: A numeric scalar; equal to the percentage of how many rows were trimmed.
}
}
\description{
Helper function for \code{trim_data()}. Removes ("trims") all rows that have a value of 1 for any of the covariates
passed through in \code{na_covariates}.
}
